Fix a bug in tuple implementation
The buggy design uses timestamp = 0 iff a read is a fast, to save one boolean
value. However, this design fails to distinguish between a fast-read and a
slow-read whose latest version happens to be the initial version. Note that even
though transaction IDs start from 1, the initial version actually has a
timestamp 0.