Optimize performance using CV
A common problem in both tulip and paxos is re-sending request too eagerly,
which creates the problem of flooding the network when the retry interval is
much smaller than the network delay. This can be addressed by waiting on a CV
with a large enough timeout, and wake up only on specific occasions, as
described below.

For paxos, re-sending an accept request is triggered when new commands are added
to the log (if required, we can further optimize for throughput by wakingup a CV
after certain amount of entries), or after a heartbeat timeout. For tulip,
sending the slow-path prepare request should happen immediately after the
transaction validates on some classic quorum of replicas, or some retry timeout.