Revert performance tweaks for eval
The previous commit is the one used in eval

A few notes about the performance issues:

1. Finding the Paxos leader: The current design of changing the Paxos leader on
failing to submit a commit or abort can sometimes lead to changing too
frequently. Right now whenever a send fails, the group coordinator would always
change the leader.  This means that if there is a spike in the network delay all
pending sends would attempt to change the leader. A better design might be
changing the leader only every few seconds if some sends have failed multiple
times; also, include the leader in the message (but beware of outdated
messages).

2. Parallelize request session: Currently replica handles requests from the same
session sequentially (in @rp.RequestSession). Better design is to invoke those
handler and sending functions in a separate Go routines.

3. Spawning too many Go routines: Sometimes running two groups on a 16-vCPU
instance using 16 clients (reaching about 30 K txns / second, with each txn
writing 4 keys) over a datacenter network can cause the replica server to be
killed by the OS. My guess is that the current design of Paxos creates too many
Go routines, but it requires more investigation.